Remove poolboy
Poolboy was originally included because we were starting a shared type
registry for workers. We needed a supervisor to create and populate the
OT types table before workers were started. Now that types are populated
in the `:ot_server` application config, the supervisor (and thus poolboy
at the library level) are no longer needed.

Users would still be encouraged to use poolboy, however, in their own
applications when using this module.